Moved files
Moved AlertWindows and WebcamUtils from common to lib as I said two days ago.
This is my last pull request, I promise :)
