+++ code/simple-webcam/src/main/java/com/nbicocchi/javafx/camera/lib/WebcamUtils.java
@@ -1,4 +1,4 @@
-package com.nbicocchi.javafx.camera.common;
+package com.nbicocchi.javafx.camera.lib;
 
 import java.awt.*;
 import java.util.Objects;
